% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{collectFunNames}
\alias{collectFunNames}
\title{Collect all functions in a package or directory}
\usage{
collectFunNames(x)
}
\arguments{
\item{x}{A character string specifying the path to an R package, directory, or file}
}
\value{
A named list of function assignments in each `.R` file in `x`
}
\description{
collect all the functions defined in an R program, directory, or file
}
\examples{
system.file("extdata", package = "pkgGraphR") |> 
    collectFunNames()

}
